started to implement instructions

+++ packages/shared/src/emulators/game_boy_color/memory_map.rs
@@ -9,29 +9,29 @@ impl MemoryMap {
     }
   }
 
-  pub fn read_u8(&self, index: usize) -> u8 {
-    self.buffer[index]
+  pub fn read_u8(&self, index: u16) -> u8 {
+    self.buffer[index as usize]
   }
 
-  pub fn read_u16(&self, index: usize) -> u16 {
-    (self.buffer[index + 0] as u16) << 0 | (self.buffer[index + 1] as u16) << 8
+  pub fn read_u16(&self, index: u16) -> u16 {
+    (self.buffer[index as usize + 0] as u16) << 0 | (self.buffer[index as usize + 1] as u16) << 8
   }
 
-  pub fn read_i8(&self, index: usize) -> i8 {
-    self.buffer[index] as i8
+  pub fn read_i8(&self, index: u16) -> i8 {
+    self.buffer[index as usize] as i8
   }
 
-  pub fn write(&mut self, index: usize, value: &[u8]) {
-    (&mut self.buffer[index..(index + value.len())]).copy_from_slice(value);
+  pub fn write(&mut self, index: u16, value: &[u8]) {
+    (&mut self.buffer[index as usize..(index as usize + value.len())]).copy_from_slice(value);
   }
 
-  pub fn write_u8(&mut self, index: usize, value: u8) {
-    self.buffer[index] = value;
+  pub fn write_u8(&mut self, index: u16, value: u8) {
+    self.buffer[index as usize] = value;
   }
 
-  pub fn write_u16(&mut self, index: usize, value: u16) {
-    self.buffer[index + 0] = (value & 0xFF) as u8;
-    self.buffer[index + 1] = (value >> 8) as u8;
+  pub fn write_u16(&mut self, index: u16, value: u16) {
+    self.buffer[index as usize + 0] = (value & 0xFF) as u8;
+    self.buffer[index as usize + 1] = (value >> 8) as u8;
   }
 
   // pub fn dump_memory_map(&self, from: usize, to: usize) {
